Remove hard coded “Delete” + `<Warning>` component
